1.14.4 #739 fix beacon rendering issue
